Validation: Strip unkown + Express/Auth added security, updated resource actions

+++ src/core/domains/express/actions/resourceDelete.ts
@@ -1,6 +1,10 @@
 import Repository from '@src/core/base/Repository';
+import ForbiddenResourceError from '@src/core/domains/auth/exceptions/ForbiddenResourceError';
+import { SecurityIdentifiers } from '@src/core/domains/auth/services/Security';
+import SecurityReader from '@src/core/domains/auth/services/SecurityReader';
 import { IRouteResourceOptions } from '@src/core/domains/express/interfaces/IRouteResourceOptions';
 import responseError from '@src/core/domains/express/requests/responseError';
+import { RouteResourceTypes } from '@src/core/domains/express/routing/RouteResource';
 import { BaseRequest } from "@src/core/domains/express/types/BaseRequest.t";
 import ModelNotFound from '@src/core/exceptions/ModelNotFound';
 import { Response } from 'express';
@@ -15,12 +19,19 @@ import { Response } from 'express';
  */
 export default async (req: BaseRequest, res: Response, options: IRouteResourceOptions): Promise<void> => {
     try {
+        const resourceOwnerSecurity = SecurityReader.findFromRouteResourceOptions(options, SecurityIdentifiers.RESOURCE_OWNER, RouteResourceTypes.DESTROY);
+
         const repository = new Repository(options.resource);
 
         const result = await repository.findById(req.params?.id);
 
         if (!result) {
             throw new ModelNotFound('Resource not found');
+        }
+
+        if(resourceOwnerSecurity && !resourceOwnerSecurity.callback(result)) {
+            responseError(req, res, new ForbiddenResourceError(), 401)
+            return;
         }
 
         await result.delete();
@@ -30,9 +41,11 @@ export default async (req: BaseRequest, res: Response, options: IRouteResourceOp
     catch (err) {
         if(err instanceof ModelNotFound) {
             responseError(req, res, err, 404)
+            return;
         }
         if (err instanceof Error) {
             responseError(req, res, err)
+            return;
         }
 
         res.status(500).send({ error: 'Something went wrong' })
